== Big Fat Refactor Update ==
* Update of the MessageObject Protocol Buffer. No more "User" message, just user id within the MessageObject message.
* Created serializable User for Client and Server. This User will be created in the client and send to the server once the client logs in. When another user logs in, all online users will be send to the new client, only once.
* Refactored Clienthandler to handle only what a client needs to handle on the server side.
* Refactored Server so that this will handle sending messages instead of the Clienthandler. Clienthandler will give the recipients id and Server finds what user to send the message to.
* Added a client handler listener to handle the functionality we just talked about above.
* Server side only works with ID's now. Message ID's or User ID's.
* Added user Status enum, so that a user can change its status. An Integer will be send, that corresponds to a Status, to others so they now what status the other client has.

+++ client/src/main/java/com/wisekrakr/wisesecurecomm/App.java
@@ -1,7 +1,7 @@
 package com.wisekrakr.wisesecurecomm;
 
 import com.wisekrakr.wisesecurecomm.communication.proto.MessageObject;
-import com.wisekrakr.wisesecurecomm.communication.proto.User;
+import com.wisekrakr.wisesecurecomm.communication.user.User;
 import com.wisekrakr.wisesecurecomm.fx.events.EventManager;
 import com.wisekrakr.wisesecurecomm.fx.screens.traynotifications.TrayNotificationType;
 
@@ -66,7 +66,7 @@ public class App implements Serializable {
                     }
 
                     @Override
-                    public void onGetOnlineUser(Map<Integer, User> users, User activeUser) {
+                    public void onGetOnlineUser(Map<Long, User> users, User activeUser) {
                         eventManager.chatAPI().getUsersOnline(users, activeUser);
                     }
